# swagger-api-contract.yml
openapi: 3.0.0
info:
  title: SmartEat Microservices API
  description: |
    Complete API documentation for the SmartEat application microservices.
    This includes User Service, Restaurant Service, Order Service, and Notification Service.
  version: 1.0.0
  contact:
    name: Development Team
    email: dev@smarteat.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:1000/api
    description: User Service (Local Development)
  - url: http://localhost:4000/api
    description: Restaurant Service (Local Development)
  - url: http://localhost:3000/api
    description: Order Service (Local Development)
  - url: http://localhost:2000/api
    description: Notification Service (Local Development)
  - url: https://user-app.wittyglacier-d214fa2c.southeastasia.azurecontainerapps.io/api
    description: User Service (Production Azure)
  - url: https://restaurant-app.wittyglacier-d214fa2c.southeastasia.azurecontainerapps.io/api
    description: Restaurant Service (Production Azure)
  - url: https://order-app.wittyglacier-d214fa2c.southeastasia.azurecontainerapps.io/api
    description: Order Service (Production Azure)
  - url: https://notification-app.wittyglacier-d214fa2c.southeastasia.azurecontainerapps.io/api
    description: Notification Service (Production Azure)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token required for protected endpoints
  schemas:
...
    Restaurant:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439022"
        name:
          type: string
          example: "Pasta Italia"
        address:
          type: string
          example: "123 Main Street, City"
        cuisine:
          type: string
          example: "Italian"
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          example: 4.5
        image:
          type: string
          format: uri
          example: "https://example.com/restaurant.jpg"
        ownerId:
          type: string
          description: ID of the restaurant owner
          example: "507f1f77bcf86cd799439011"
        createdAt:
          type: string
          format: date-time
          example: "2024-03-20T10:30:00Z"
    CreateRestaurantRequest:
      type: object
      properties:
        name:
          type: string
          example: "Pasta Italia"
        address:
          type: string
          example: "123 Main Street, City"
        cuisine:
          type: string
          example: "Italian"
        rating:
          type: number
          format: float
          example: 4.5
        image:
          type: string
          format: uri
          example: "https://example.com/restaurant.jpg"
      required:
        - name
        - address
        - cuisine
    MenuItem:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439033"
        restaurantId:
          type: string
          example: "507f1f77bcf86cd799439022"
        name:
          type: string
          example: "Spaghetti Carbonara"
        description:
          type: string
          example: "Traditional Italian pasta with bacon and cream sauce"
        price:
          type: number
          format: float
          minimum: 0
          example: 12.99
        image:
          type: string
          format: uri
          example: "https://example.com/dish.jpg"
    CreateMenuItemRequest:
      type: object
      properties:
        name:
          type: string
          example: "Spaghetti Carbonara"
        description:
          type: string
          example: "Traditional Italian pasta with bacon and cream sauce"
        price:
          type: number
          format: float
          example: 12.99
        image:
          type: string
          format: uri
          example: "https://example.com/dish.jpg"
      required:
        - name
        - description
        - price
    UpdateMenuItemRequest:
      type: object
      properties:
        name:
          type: string
          example: "Spaghetti Carbonara"
        description:
          type: string
          example: "Traditional Italian pasta with bacon and cream sauce"
        price:
          type: number
          format: float
          example: 12.99
        image:
          type: string
          format: uri
          example: "https://example.com/dish.jpg"
...
    Notification:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439055"
        userId:
          type: string
          example: "507f1f77bcf86cd799439011"
        type:
          type: string
          enum: [order_update, order_delivered, order_cancelled, payment_confirmed]
          example: "order_update"
        message:
          type: string
          example: "Your order has been confirmed"
        orderId:
          type: string
          example: "507f1f77bcf86cd799439044"
        read:
          type: boolean
          default: false
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2024-03-20T10:30:00Z"
    CreateNotificationRequest:
      type: object
      properties:
        userId:
          type: string
          example: "507f1f77bcf86cd799439011"
        type:
          type: string
          enum: [order_update, order_delivered, order_cancelled, payment_confirmed]
          example: "order_update"
        message:
          type: string
          example: "Your order has been confirmed"
        orderId:
          type: string
          example: "507f1f77bcf86cd799439044"
        read:
          type: boolean
          default: false
      required:
        - userId
        - type
        - message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error message"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
paths:
...
  /restaurants:
    get:
      tags:
        - Restaurant
      summary: List all restaurants
      description: Get a list of all available restaurants
      operationId: getRestaurants
      responses:
        '200':
          description: List of restaurants
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
    post:
      tags:
        - Restaurant
      summary: Create a new restaurant
      description: Create a new restaurant (requires authentication)
      operationId: createRestaurant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRestaurantRequest'
      responses:
        '201':
          description: Restaurant created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Restaurant'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /restaurants/{id}:
    get:
      tags:
        - Restaurant
      summary: Get restaurant by ID
      description: Retrieve detailed information about a specific restaurant
      operationId: getRestaurantById
      parameters:
        - name: id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
      responses:
        '200':
          description: Restaurant details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Restaurant'
        '404':
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Restaurant
      summary: Update restaurant
      description: Update restaurant details (requires authentication)
      operationId: updateRestaurant
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRestaurantRequest'
      responses:
        '200':
          description: Restaurant updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Restaurant'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Restaurant
      summary: Delete restaurant
      description: Delete a restaurant (requires authentication)
      operationId: deleteRestaurant
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
      responses:
        '200':
          description: Restaurant deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /restaurants/owner:
    get:
      tags:
        - Restaurant
      summary: Get owner's restaurants
      description: Get list of restaurants owned by authenticated user
      operationId: getOwnerRestaurants
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of owner's restaurants
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /restaurants/{id}/menu:
    get:
      tags:
        - Menu
      summary: Get restaurant menu
      description: Get all menu items for a specific restaurant
      operationId: getMenuByRestaurant
      parameters:
        - name: id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
      responses:
        '200':
          description: List of menu items
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MenuItem'
        '404':
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Menu
      summary: Create menu item
      description: Add a new menu item to a restaurant
      operationId: createMenuItem
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuItemRequest'
      responses:
        '201':
          description: Menu item created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/MenuItem'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /restaurants/{id}/menu/{menuId}:
    put:
      tags:
        - Menu
      summary: Update menu item
      description: Update details of a menu item
      operationId: updateMenuItem
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
        - name: menuId
          in: path
          required: true
          description: Menu Item ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMenuItemRequest'
      responses:
        '200':
          description: Menu item updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/MenuItem'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Menu item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Menu
      summary: Delete menu item
      description: Remove a menu item from a restaurant
      operationId: deleteMenuItem
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: string
        - name: menuId
          in: path
          required: true
          description: Menu Item ID
          schema:
            type: string
      responses:
        '200':
          description: Menu item deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Menu item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /notifications:
    post:
      tags:
        - Notification
      summary: Create notification
      description: Create a new notification (usually called by services)
      operationId: createNotification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationRequest'
      responses:
        '201':
          description: Notification created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Notification'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...